(develop) Fixing an assortment of bugs and warnings found when tracking down issue for failed autotest on linux (test_filter_second.glm): Fixed a memory overrun on array in dishwasher.cpp Fixed an improper method for comparing strings in inverter.cpp Fixed an overflow of int in substation.cpp Fixed a warning in regulator.cpp by changing char to int Fixed what looks to be copy paste errors along with improper order of operations in shedule.c Fixed an error related to min in job.cpp and validate.cpp by changing to fmin  (to comply with clang 9.0.0) Fixed improper comparison between int and pointer in csv_reader.cpp (to comply with clang 9.0.0)

+++ powerflow/substation.cpp
@@ -389,7 +389,7 @@ SIMULATIONMODE substation::inter_deltaupdate_substation(unsigned int64 delta_tim
 		//calculate the energy used
 		if(iteration_count_val == 0){
 			total_load = last_power_A.Re() + last_power_B.Re() + last_power_C.Re();
-			distribution_real_energy += total_load*dt/(3600*DT_SECOND);
+			distribution_real_energy += total_load*((double)dt/(3600.0*DT_SECOND));
 		}
 		NR_node_presync_fxn(0);
